requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
tavily-python>=0.3.0
pydantic>=2.0.0
openai>=1.0.0
